test suite: move breakout room API tests to a separate file
